# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/demandes/{id}/taches": {
      "get": {
        "operationId": "DemandesController_listerTaches",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "demandes"
        ]
      }
    },
...
    "/api/demandes/{id}/calcul": {
      "post": {
        "operationId": "DemandesController_recalculer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "demande": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "reference": {
                          "type": "string"
                        },
                        "circuit": {
                          "type": "string",
                          "enum": [
                            "DOBB",
                            "DXC",
                            "DF"
                          ]
                        },
                        "segment": {
                          "type": "string"
                        },
                        "sousFlux": {
                          "type": "string",
                          "nullable": true
                        },
                        "nomClient": {
                          "type": "string"
                        },
                        "compteClient": {
                          "type": "string",
                          "nullable": true
                        },
                        "agentInitiateur": {
                          "type": "string",
                          "nullable": true
                        },
                        "matriculeInitiateur": {
                          "type": "string",
                          "nullable": true
                        },
                        "agentSaisie": {
                          "type": "string",
                          "nullable": true
                        },
                        "localisation": {
                          "type": "string",
                          "enum": [
                            "NATIONAL",
                            "INTERNATIONAL"
                          ],
                          "nullable": true
                        },
                        "canalRemontee": {
                          "type": "string",
                          "nullable": true
                        },
                        "dateReceptionBo": {
                          "type": "string",
                          "nullable": true
                        },
                        "dateReceptionOci": {
                          "type": "string",
                          "nullable": true
                        },
                        "formuleAbonnement": {
                          "type": "string",
                          "nullable": true
                        },
                        "numeroAppel": {
                          "type": "string",
                          "nullable": true
                        },
                        "debutPeriodeContestee": {
                          "type": "string",
                          "nullable": true
                        },
                        "finPeriodeContestee": {
                          "type": "string",
                          "nullable": true
                        },
                        "periodeContesteeJours": {
                          "type": "number",
                          "nullable": true
                        },
                        "recurrentMensuel": {
                          "type": "boolean"
                        },
                        "champsCircuit": {
                          "type": "object",
                          "additionalProperties": {}
                        },
                        "montantHt": {
                          "type": "number"
                        },
                        "montantTsc": {
                          "type": "number"
                        },
                        "montantTva": {
                          "type": "number"
                        },
                        "montantTtc": {
                          "type": "number"
                        },
                        "tscActive": {
                          "type": "boolean"
                        },
                        "tvaActive": {
                          "type": "boolean"
                        },
                        "tauxTsc": {
                          "type": "number"
                        },
                        "tauxTva": {
                          "type": "number"
                        },
                        "libelle": {
                          "type": "string",
                          "nullable": true
                        },
                        "motifId": {
                          "type": "string",
                          "format": "uuid",
                          "nullable": true
                        },
                        "universFmiCode": {
                          "type": "string",
                          "nullable": true
                        },
                        "facteurCode": {
                          "type": "string",
                          "nullable": true
                        },
                        "directionRespId": {
                          "type": "string",
                          "format": "uuid",
                          "nullable": true
                        },
                        "serviceRespId": {
                          "type": "string",
                          "format": "uuid",
                          "nullable": true
                        },
                        "agentResponsable": {
                          "type": "string",
                          "nullable": true
                        },
                        "statut": {
                          "type": "string",
                          "enum": [
                            "BROUILLON",
                            "SOUMIS",
                            "EN_COURS",
                            "VALIDE",
                            "REJETE",
                            "ABANDONNE"
                          ]
                        },
                        "etapeCourante": {
                          "type": "number"
                        },
                        "initiateurId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "dateDemande": {
                          "type": "string"
                        },
                        "dateSoumission": {
                          "type": "string",
                          "nullable": true
                        },
                        "dateCloture": {
                          "type": "string",
                          "nullable": true
                        },
                        "commentaire": {
                          "type": "string",
                          "nullable": true
                        },
                        "responsabiliteServiceAutre": {
                          "type": "string",
                          "nullable": true
                        },
                        "siEtat": {
                          "type": "string",
                          "enum": [
                            "EN_ATTENTE",
                            "ENVOYE",
                            "CONFIRME",
                            "ERREUR"
                          ]
                        },
                        "siRef": {
                          "type": "string",
                          "nullable": true
                        },
                        "siHorodatage": {
                          "type": "string",
                          "nullable": true
                        },
                        "siMessage": {
                          "type": "string",
                          "nullable": true
                        },
                        "siTentatives": {
                          "type": "number"
                        },
                        "siAdaptateur": {
                          "type": "string",
                          "nullable": true
                        }
                      },
                      "required": [
                        "id",
                        "reference",
                        "circuit",
                        "segment",
                        "sousFlux",
                        "nomClient",
                        "compteClient",
                        "agentInitiateur",
                        "matriculeInitiateur",
                        "agentSaisie",
                        "localisation",
                        "canalRemontee",
                        "dateReceptionBo",
                        "dateReceptionOci",
                        "formuleAbonnement",
                        "numeroAppel",
                        "debutPeriodeContestee",
                        "finPeriodeContestee",
                        "periodeContesteeJours",
                        "recurrentMensuel",
                        "champsCircuit",
                        "montantHt",
                        "montantTsc",
                        "montantTva",
                        "montantTtc",
                        "tscActive",
                        "tvaActive",
                        "tauxTsc",
                        "tauxTva",
                        "libelle",
                        "motifId",
                        "universFmiCode",
                        "facteurCode",
                        "directionRespId",
                        "serviceRespId",
                        "agentResponsable",
                        "statut",
                        "etapeCourante",
                        "initiateurId",
                        "dateDemande",
                        "dateSoumission",
                        "dateCloture",
                        "commentaire",
                        "responsabiliteServiceAutre",
                        "siEtat",
                        "siRef",
                        "siHorodatage",
                        "siMessage",
                        "siTentatives",
                        "siAdaptateur"
                      ],
                      "additionalProperties": false
                    },
                    "lignes": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "ligneId": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "nd": {
                            "type": "string"
                          },
                          "formuleId": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "recurrent": {
                            "type": "number"
                          },
                          "recurrentModifie": {
                            "type": "boolean"
                          },
                          "statutLigne": {
                            "type": "string",
                            "enum": [
                              "ACTIF",
                              "SUSPENDU",
                              "RESILIE"
                            ]
                          },
                          "montantHtLigne": {
                            "type": "number"
                          },
                          "debutPeriodeContestee": {
                            "type": "string",
                            "nullable": true
                          },
                          "finPeriodeContestee": {
                            "type": "string",
                            "nullable": true
                          },
                          "periodeContesteeJours": {
                            "type": "number",
                            "nullable": true
                          }
                        },
                        "required": [
                          "id",
                          "ligneId",
                          "nd",
                          "formuleId",
                          "recurrent",
                          "recurrentModifie",
                          "statutLigne",
                          "montantHtLigne",
                          "debutPeriodeContestee",
                          "finPeriodeContestee",
                          "periodeContesteeJours"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "pieces": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "pieceAfferenteId": {
                            "type": "string",
                            "format": "uuid",
                            "nullable": true
                          },
                          "nomFichier": {
                            "type": "string"
                          },
                          "typeMime": {
                            "type": "string"
                          },
                          "tailleOctets": {
                            "type": "number"
                          },
                          "gedRef": {
                            "type": "string",
                            "nullable": true
                          },
                          "dateAjout": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "pieceAfferenteId",
                          "nomFichier",
                          "typeMime",
                          "tailleOctets",
                          "gedRef",
                          "dateAjout"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "demande",
                    "lignes",
                    "pieces"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "tags": [
          "demandes"
        ]
      }
    },
    "/api/demandes/{id}/apercu-routage": {
      "post": {
        "operationId": "DemandesController_apercuRoutage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "labelPalier": {
                      "type": "string",
                      "nullable": true
                    },
                    "etapes": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "ordre": {
                            "type": "number"
                          },
                          "roleCode": {
                            "type": "string"
                          },
                          "typeActeur": {
                            "type": "string"
                          },
                          "bloquant": {
                            "type": "boolean"
                          },
                          "slaHeures": {
                            "type": "number"
                          }
                        },
                        "required": [
                          "ordre",
                          "roleCode",
                          "typeActeur",
                          "bloquant",
                          "slaHeures"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "labelPalier",
                    "etapes"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "tags": [
          "demandes"
        ]
      }
    },
    "/api/demandes/{id}/soumettre": {
      "post": {
        "operationId": "DemandesController_soumettre",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statut": {
                      "type": "string",
                      "enum": [
                        "BROUILLON",
                        "SOUMIS",
                        "EN_COURS",
                        "VALIDE",
                        "REJETE",
                        "ABANDONNE"
                      ]
                    },
                    "etapeCourante": {
                      "type": "number"
                    },
                    "dateSoumission": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "statut",
                    "etapeCourante",
                    "dateSoumission"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "tags": [
          "demandes"
        ]
      }
    },
    "/api/demandes/{id}/abandonner": {
      "post": {
        "operationId": "DemandesController_abandonner",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "demandes"
        ]
      }
    },
    "/api/demandes/{id}/rappeler": {
      "post": {
        "operationId": "DemandesController_rappeler",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "demandes"
        ]
      }
    },
    "/api/demandes/{id}/pieces": {
      "post": {
        "operationId": "DemandesController_ajouterPiece",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "pieceAfferenteId": {
                      "type": "string",
                      "format": "uuid",
                      "nullable": true
                    },
                    "nomFichier": {
                      "type": "string"
                    },
                    "typeMime": {
                      "type": "string"
                    },
                    "tailleOctets": {
                      "type": "number"
                    },
                    "gedRef": {
                      "type": "string",
                      "nullable": true
                    },
                    "dateAjout": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "pieceAfferenteId",
                    "nomFichier",
                    "typeMime",
                    "tailleOctets",
                    "gedRef",
                    "dateAjout"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "tags": [
          "demandes"
        ]
      }
    },
...
  },
  "info": {
    "title": "PGD — API",
    "description": "Plateforme de Gestion des Dégrèvements — Orange Côte d'Ivoire",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
...
}